Power save stop watch added
In order to save power for long time stop watch modes i implemented a
seconds accurate Mode. This one only count on ticks of the RTC and does
not use a internal timer. I still increments the stpwatch variable, so
easy switching between modes is possible but might be unwanted.